# theftp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Observatorio de Transporte Público API",
        "description": "Backend y API Acceso a datos.",
        "contact": {
            "email": "juan-vega6@upc.edu.co"
        },
        "license": {
            "name": "@Universidad Piloto de Colombia SAM - Todos los Derechos Reservados"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auditoria/{field}/uniques": {
            "get": {
                "tags": [
                    "Auditoría"
                ],
                "summary": "Obtener valores únicos de un campo en la tabla de auditoría",
                "operationId": "3b67f1b015d0e15670b0122a9a75b2ec",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "Campo de la tabla de auditoría para obtener valores únicos",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valores únicos obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "El campo no es válido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auditoria": {
            "get": {
                "tags": [
                    "Auditoría"
                ],
                "summary": "Obtener registros de auditoría paginados",
                "operationId": "ef113beb78e2fcb3120b096924850f20",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "itemsPerPage",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Texto de búsqueda",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "updated"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Columna para ordenar",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "id"
                        }
                    },
                    {
                        "name": "orderDirection",
                        "in": "query",
                        "description": "Dirección de ordenamiento (asc o desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registros de auditoría obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "itemsPerPage": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en los parámetros de entrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum Token Bearer",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}